backend/requirements.txt
Django ~=4.1.1
psycopg2 ~=2.9.1
django-cors-headers ~=3.13.0
djangorestframework ~=3.13.1
pytest ~=7.1.3
pytest-django ~=4.5.2
factory-boy ~=3.2.1
django-filter ~=22.1
coverage ~=6.4.4
flake8 ~=5.0.4
djangorestframework-simplejwt ~=5.2.0
django-extensions ~=3.2.1